"Add optional parameter to reload function in pagination tables"
In this commit, an optional parameter 'resetPage' has been added to the 'reload' function in 'AppPaginationTable.vue' and 'AppTable.vue'. This allows more flexibility by providing an option to reset the page to the first one when reloading. The change makes pagination more versatile in different scenarios where maintaining the current page number can be necessary.